MGDOBR-865: integrate smoke tests
There was created a new folder which name is 'mocked'. It contains all test suites which are not suggested to run with dev env. This folder is only temporary. All tests should be moved under the 'integration' folder.

The Util file was moved outside of the test space.
