+++ cypress/tsconfig.json
@@ -8,6 +8,8 @@
     "support/**/*.ts",
     "integration/**/*.ts",
     "e2e/**/*.ts",
+    "mocked/**/*.ts",
+    "utils/**/*.ts",
     "plugin/**/*.ts"
   ]
 }
